+++ tests/testthat/Testbox/app/modules/module.R
@@ -1,0 +1,14 @@
+#' an example function
+#'
+#' @export
+a <- function(x) {
+  if (x <= 1) {
+    1
+  } else {
+    2
+  }
+}
+
+private_function <- function(x) {
+  x ^ 2
+}