+++ tests/testthat/Testbox_attached_modules_functions/app/app.R
@@ -1,0 +1,8 @@
+options(box.path = file.path(getwd()))
+# remove box cache
+loaded_mods <- loadNamespace("box")$loaded_mods
+rm(list = ls(loaded_mods), envir = loaded_mods)
+
+box::use(
+  app/modules/module
+)