+++ NEWS.md
@@ -1,5 +1,7 @@
 # covr (development version)
 
+* Added support for `klmr/box` modules. This works best with `file_coverage()`. (@radbasa, #491)
+
 # covr 3.6.4
 
 * Fix for a failing test on CRAN